+++ src/constants/titles.js
@@ -1,0 +1,4 @@
+// Title prefixes for each part of the application
+export const OFFICE_BASE_PAGE_TITLE = 'Transcom PPP: Office';
+export const ADMIN_BASE_PAGE_TITLE = 'Transcom PPP: Admin';
+export const MILMOVE_BASE_PAGE_TITLE = 'my.move.mil';
