+++ src/scenes/MyMove/index.jsx
@@ -20,7 +20,8 @@ import LoadingPlaceholder from 'shared/LoadingPlaceholder';
 import SomethingWentWrong from 'shared/SomethingWentWrong';
 import { loadInternalSchema } from 'shared/Swagger/ducks';
 import { withContext } from 'shared/AppContext';
-import { generatePageTitle, no_op } from 'shared/utils';
+import { no_op } from 'shared/utils';
+import { generatePageTitle } from 'hooks/custom';
 import { loadUser as loadUserAction } from 'store/auth/actions';
 import { initOnboarding as initOnboardingAction } from 'store/onboarding/actions';
 import { selectGetCurrentUserIsLoading, selectIsLoggedIn } from 'store/auth/selectors';
